# TrackIt.Frontend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TrackIt API",
    "version": "v1"
  },
  "paths": {
...
    "/api/transactions": {
      "post": {
        "tags": [
          "Transaction"
        ],
        "summary": "Создание транзакции",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionCommand"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Transaction"
        ],
        "summary": "Получение всех транзакций",
        "parameters": [
          {
            "name": "category_id",
            "in": "query",
            "description": "Идентификатор категории",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Номер страницы",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Предел",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDtoPagedList"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDtoPagedList"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDtoPagedList"
                }
              }
            }
          }
        }
      }
    },
    "/api/transactions/{id}": {
      "put": {
        "tags": [
          "Transaction"
        ],
        "summary": "Обновление транзакции",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Transaction"
        ],
        "summary": "Удаление транзакции",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Transaction"
        ],
        "summary": "Получение транзакции по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DetailedTransactionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailedTransactionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailedTransactionDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTransactionCommand": {
        "type": "object",
        "properties": {
          "category_id": {
            "type": "string",
            "description": "Идентификатор категории.",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "description": "Сумма транзакции.",
            "format": "double"
          },
          "description": {
            "type": "string",
            "description": "Описание транзакции.",
            "nullable": true
          },
          "date": {
            "type": "string",
            "description": "Дата транзакции.",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Команда для создания новой транзакции."
      },
...
      "DetailedTransactionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор транзакции.",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "description": "Уникальный идентификатор пользователя.",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "description": "Сумма транзакции.",
            "format": "double"
          },
          "date": {
            "type": "string",
            "description": "Дата транзакции.",
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "description": "Дата создания",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "description": "Дата обновления",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "description": "Описание",
            "nullable": true
          },
          "category_id": {
            "type": "string",
            "description": "Идентификатор категории",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Детализированная модель транзакции"
      },
...
      "TransactionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор транзакции.",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "description": "Уникальный идентификатор пользователя.",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "description": "Сумма транзакции.",
            "format": "double"
          },
          "date": {
            "type": "string",
            "description": "Дата транзакции.",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Модель транзакции"
      },
      "TransactionDtoPagedList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionDto"
            },
            "description": "Элементы",
            "nullable": true
          },
          "total": {
            "type": "integer",
            "description": "Всего элементов",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Пагинированный список"
      },
...
      "UpdateTransactionCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор транзакции.",
            "format": "uuid"
          },
          "category_id": {
            "type": "string",
            "description": "Уникальный идентификатор категории.",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "description": "Сумма транзакции.",
            "format": "double"
          },
          "description": {
            "type": "string",
            "description": "Описание транзакции.",
            "nullable": true
          },
          "date": {
            "type": "string",
            "description": "Дата транзакции.",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Команда для обновления транзакции."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Введите ваш JWT-токен в формате: Bearer {token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}